restructure source, add controller/worker setup
