+++ done_soon/data_generation/db/__init__.py
@@ -1,0 +1,2 @@
+from . import datastructs
+from . import functions
